+++ src/Aardvark.Base/Geometry/Types/Plane/Plane3_auto.cs
@@ -830,8 +830,9 @@ namespace Aardvark.Base
         {
             if (pointArray == null) throw new ArgumentNullException();
             if (startIndex < 0 || startIndex >= pointArray.Length) throw new ArgumentOutOfRangeException();
-            if (count < 0 || startIndex + count >= pointArray.Length) throw new ArgumentOutOfRangeException();
+            if (count < 0) throw new ArgumentOutOfRangeException();
             if (count == 0) count = pointArray.Length - startIndex;
+            if (count > pointArray.Length - startIndex) throw new ArgumentOutOfRangeException();
 
             var normal = plane.Normal;
             var result = new V3f[count];
@@ -851,8 +852,9 @@ namespace Aardvark.Base
         {
             if (pointArray == null) throw new ArgumentNullException();
             if (startIndex < 0 || startIndex >= pointArray.Length) throw new ArgumentOutOfRangeException();
-            if (count < 0 || startIndex + count >= pointArray.Length) throw new ArgumentOutOfRangeException();
+            if (count < 0) throw new ArgumentOutOfRangeException();
             if (count == 0) count = pointArray.Length - startIndex;
+            if (count > pointArray.Length - startIndex) throw new ArgumentOutOfRangeException();
 
             var result = new V3f[count];
             for (int i = startIndex, j = 0; j < count; i++, j++)
@@ -1741,8 +1743,9 @@ namespace Aardvark.Base
         {
             if (pointArray == null) throw new ArgumentNullException();
             if (startIndex < 0 || startIndex >= pointArray.Length) throw new ArgumentOutOfRangeException();
-            if (count < 0 || startIndex + count >= pointArray.Length) throw new ArgumentOutOfRangeException();
+            if (count < 0) throw new ArgumentOutOfRangeException();
             if (count == 0) count = pointArray.Length - startIndex;
+            if (count > pointArray.Length - startIndex) throw new ArgumentOutOfRangeException();
 
             var normal = plane.Normal;
             var result = new V3d[count];
@@ -1762,8 +1765,9 @@ namespace Aardvark.Base
         {
             if (pointArray == null) throw new ArgumentNullException();
             if (startIndex < 0 || startIndex >= pointArray.Length) throw new ArgumentOutOfRangeException();
-            if (count < 0 || startIndex + count >= pointArray.Length) throw new ArgumentOutOfRangeException();
+            if (count < 0) throw new ArgumentOutOfRangeException();
             if (count == 0) count = pointArray.Length - startIndex;
+            if (count > pointArray.Length - startIndex) throw new ArgumentOutOfRangeException();
 
             var result = new V3d[count];
             for (int i = startIndex, j = 0; j < count; i++, j++)